+++ CHANGELOG.md
@@ -9,6 +9,9 @@ Changelog
 * Disable reporting of notable addresses from registers for handled errors
   [#741](https://github.com/bugsnag/bugsnag-cocoa/pull/741)
 
+* Improve thread-safety of breadcrumb count
+  [#743](https://github.com/bugsnag/bugsnag-cocoa/pull/743)
+
 ## 6.0.1 (2020-06-29)
 
 * Move binary images store declaration from header file